feat: revamp next.js preview
