+++ app/constants.ts
@@ -1,8 +1,8 @@
-import { FlaskConical, Home, MousePointerSquareDashed } from "lucide-react";
+import { Home, MousePointerSquareDashed, Settings } from "lucide-react";
 import type { Tabs } from "../src/layout/TabBar/TabBarItem";
 
 export const TABS: Tabs = [
   { text: "Home", href: "/", icon: Home },
-  { text: "Preview", href: "/preview", icon: FlaskConical },
+  { text: "Settings", href: "/settings", icon: Settings },
   { text: "Ladle", href: "/ladle", icon: MousePointerSquareDashed },
 ];
